requirements.txt
arrow==0.15.5
asgiref==3.2.3
blessed==1.16.1
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.2
django-filter==2.2.0
django-picklefield==2.0
django-q==1.0.2
djangorestframework==3.11.0
drf-yasg==1.17.0
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.9
idna==2.8
inflection==0.3.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
openpyxl==3.0.2
packaging==20.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.8
psycopg2==2.8.4
gunicorn
meinheld
environs